fix: 매니지먼트 모듈 리팩토링중
- 지수 응답 시, 백분율 계산 로직 필요
- 로그인 시, exp 부여 (1회) 로직 필요
- lifecycle 모듈 지수 감소에 따른 상태(state) 변경 로직 필요
- FeedHistory 기능 부분 단위 테스트 필요
- MongActive 는 로그 생성용으로 사용
- Enum 을 통해 로직의 경계값(ex. 진화 가능 exp 등)으로 사용할 수 있도록 로직 수정 필요

+++ management/src/main/java/com/mongs/management/domain/mong/controller/dto/response/FindMongResDto.java
@@ -17,7 +17,7 @@ public record FindMongResDto(
         Double sleep,
         Integer poopCount,
         Boolean isSleeping,
-        Integer exp,
+        Double exp,
         LocalDateTime born,
         String stateCode,
         String shiftCode,
