+++ management/src/main/java/com/mongs/management/domain/mong/controller/dto/response/RegisterMongResDto.java
@@ -16,7 +16,7 @@ public record RegisterMongResDto(
         Double sleep,
         Integer poopCount,
         Boolean isSleeping,
-        Integer exp,
+        Double exp,
         LocalDateTime born,
         String stateCode,
         String shiftCode,
